func(sdk2.0): sdk 2.0

+++ tcvdb_text/src/main/java/com/tencent/tcvectordb/hash/BaseHash.java
@@ -1,5 +1,5 @@
 package com.tencent.tcvectordb.hash;
 
 public interface BaseHash {
-    public int hash(String data);
+    public Long hash(String data);
 }
